Changed how animation events are fired
Its faster and easier to just pass through an event the id of the animation event and whoever wants to listen to the event checks if that id matches with his id.

Doing it this way removed the interface, the abstract class and the lists. Now any class can work as an animation event.